requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Deprecated==1.2.7
Django==3.0.2
idna==2.8
PyGithub==1.45
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
wrapt==1.11.2
